modetest: Add virtual_width and pixels_per_container
Add two variables to bo_create() to help allocating more complex pixel
formats: virtual_width and pixels_per_container.

virtual_width is width divided by pixels_per_container. By default (for
all current formats) pixels_per_container is 1, which means that the bpp
value tells the bits-per-pixel. Thus virtual_width == width in that
case, and it's business as usual.

However, when pixels_per_container is higher than 1, the bpp value tells
the size of the container, and pixels_per_container tells the number of
pixels inside that container.

Thus when we call bo_create_dumb(), we provide it the bits-per-container
(as bpp), and width (in the number of containers).

The rationale here is that when more complex pixel packing is used, the
normal bits-per-pixel becomes unusable: e.g. Y10_P32 packs three 10-bit
pixel values into a 32-bit container. Using bpp = 10 is not correct, as
there are two padding bits. Bpp = 32/3 would be (kind of) correct, but
it is not an integer.

Thus, with this patch, we can present such formats as
pixels_per_container = 3, bpp = 32.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>